More work on hand and all over.
Simplified API calls and removed a lot of logging
Added untested AbstractGrouping.insert_transform_buffer as a way to insert a buffer node from the context of a Grouping.
Made the API property a true dictionary and allowed for APIs to be accessible along with the new APIProxy.DEFAULT_API property.

Added fileop function to dcc.scene.
Completely removed digit as it makes more sense to add as a method than a class.
Added test suite for Hand but haven't filled it in.