Started reworking the HierarchyChain and relationships.py classes in general
Refactored the names of the classes to be more demonstrative.

Trying to refactor all simple functions to the base class and simplify the classes in general.

Added simple pre/post scene sanitization wrapper based on cleanup_nodes.  Might just build this into the test runner of the base class so I don't have to keep tagging.