Removed Bunch implementation as it was failing due to displaying data errors during printing/yaml load/dumping.
Converted generate_build_report to a wrapper.  This will usually be chained with register_built_nodes.
Removed insert_transform_buffer as it was a duplicate from NonLinearHierarchySet.

Starting to change the logic behind Control as it is acting a bit non sensically as it currently stands.

Modifying the main build process to match the new registration process.  Need to find a better plan than getting the last node created in the nodes list.

Refactored test_hierarchy_chain to test_relationships and added new tests for registration in test_abstract_grouping.

Fixed the wrapping functionality for pre_and_post_sanitize_scene with @functools.wraps.