Added anvil.utils.generic.gen_flatten_dict_depth_two for quickly flattening the hierarchy Map instance. Added anvil.grouping.AbstractGrouping._flatten_hierarchy() to follow up.
Fixed Control implementation and simplified build steps to be more legible.
Modified all build methods to classmethods, however this is breaking the SubRigTemplate builds as it is passing as a class and not an instance.  Need to figure this one out next.

Removed SubRigTemplate.get_shape_list since it was redundant to to_size_list which is much better and generalized.

Added MetaData.to_dict to stop pissing off kwargs linting.

Modified so many tests to get in line with new hierarchy registration.