Fixing errors based on CircleCI build 224 https://circleci.com/gh/AndresMWeber/Anvil/224
Minor mistakes like not pulling out the name tokens from kwargs when redefining or adding to.
Forgot to use _flat_hierarchy to pull all controls out easily to the top level for testing against.
Forgot to change dot notation access to the result dict to getitem access as it is raw and not yet a mapping.
Changed property implementation of NodeSet.ANVIL_TYPE