Added skip_register flag to skip registration if we are doing a chained build call (so things get sorted properly to the hierarchy). Removed is_class since it actually did not work.
Added testing to see if we are within CIRCLECI.
Simplified safe_delete.
Refactored cleanup_nodes to sanitize for brevity.

Removed a ton of functions from BaseTest as they are not in use anymore.

Added dict_deep_sort to convert a dictionary of dictionaries to a sorted nest of OrderedDict

Moved sanitize_scene to utils.scene.