Fixed parenting for controls and added an override_root kwarg so we can not rely on instance.root being set and can use arbitrary "top" nodes to parent a rig
Fixed the auto_save_results so that we are now properly saving out files that represent the state of the test case.  This will make testing MUCH easier and quicker.

Changed auto_save_results to read environment variables instead so that we can set env vars instead of rely on passing kwargs which is less easy to do with unittests.