Added TypeError to the API errors we check for to give a better stack trace.
Figured out that parent.pop was not working because we were pulling from the end of the list, fixed so the parenting works properly in the build methods of the SubRigTemplate

Moved the test_build_no_kwargs to a different TestCase so I can pick which test files to save easier.

Added a timestamp to the test save files so we aren't overwriting constantly.